+++ .gitignore
@@ -12,6 +12,7 @@ target/
 .settings
 .springBeans
 .sts4-cache
+/src/main/resources/client_secret.json
 
 ### IntelliJ IDEA ###
 .idea
