+++ src/main/java/br/com/felipefinger/mita/controller/PacienteResource.java
@@ -1,4 +1,4 @@
-package br.com.felipefinger.mita.resources;
+package br.com.felipefinger.mita.controller;
 
 import java.util.List;
