+++ src/main/java/br/com/felipefinger/mita/controller/VideoController.java
@@ -1,10 +1,12 @@
-package br.com.felipefinger.mita.resources;
+package br.com.felipefinger.mita.controller;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -12,16 +14,21 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import br.com.felipefinger.mita.dto.VideoDTO;
 import br.com.felipefinger.mita.models.Video;
 import br.com.felipefinger.mita.repository.VideoRepository;
+import br.com.felipefinger.mita.service.VideoService;
 
 @RestController
 @RequestMapping(value = "/videos")
-public class VideoResource {
+public class VideoController {
 
 	@Autowired
 	VideoRepository videoRepository;
 	
+	@Autowired
+	VideoService videoService;
+
 	@GetMapping("/adquirir")
 	public List<Video> adquirir(String titulo){
 		
@@ -39,13 +46,18 @@ public class VideoResource {
 		return videoRepository.findById(id);
 	}
 	
-	@PostMapping("/salvar")
-	public Video salvar(@RequestBody Video video){
+	@PostMapping(value="/salvar", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+	public Video salvar(@ModelAttribute VideoDTO videoDTO) throws Exception{
 		
-		video.setTitulo(video.getTitulo().toUpperCase());
-		return videoRepository.save(video);
+		return videoService.uploadVideo(videoDTO);
 	}
 	
+	@PostMapping("/salvarlink")
+	public Video salvarLink(@RequestBody Video video) throws Exception{
+		
+		return videoService.salvarVideo(video);
+	}
+
 	@DeleteMapping("/excluir/{id}")
 	public void excluir(@PathVariable(value = "id") long id){
