+++ src/main/java/br/com/felipefinger/mita/dto/VideoDTO.java
@@ -1,28 +1,20 @@
-package br.com.felipefinger.mita.models;
+package br.com.felipefinger.mita.dto;
 
 import java.io.Serializable;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import org.springframework.web.multipart.MultipartFile;
 
-@Entity
-@Table(name = "TB_VIDEO")
-public class Video implements Serializable {
+public class VideoDTO implements Serializable {
 
 	private static final long serialVersionUID = 1L;
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
 	private long id;
 
 	private String titulo;
 
 	private String descricao;
 
-	private String url;
+	private MultipartFile video;
 
 	public long getId() {
 		return id;
@@ -48,11 +40,15 @@ public class Video implements Serializable {
 		this.descricao = descricao;
 	}
 
-	public String getUrl() {
-		return url;
+	public MultipartFile getVideo() {
+		return video;
 	}
 
-	public void setUrl(String url) {
-		this.url = url;
+	public void setVideo(MultipartFile video) {
+		this.video = video;
+	}
+
+	public static long getSerialversionuid() {
+		return serialVersionUID;
 	}
 }
