+++ src/main/java/br/com/felipefinger/mita/service/VideoUploadService.java
@@ -1,0 +1,5 @@
+package br.com.felipefinger.mita.service;
+
+public class VideoUploadService {
+
+}
