+++ src/main/resources/application.properties
@@ -5,3 +5,5 @@ spring.datasource.url=jdbc:postgresql://localhost:5432/mita
 spring.datasource.username=postgres
 spring.datasource.password=postgres
 spring.jpa.hibernate.ddl-auto=update
+
+spring.servlet.multipart.max-file-size=-1
